Add xml to dump functionality to support full round trip
Note there are some special cases for some tags that aren't yet
correctly mapped (e.g. Row, Columns).

Also failure modes are not handled gracefully, so a bad dicom
file will abort the process.

There also appears to be a limit on the number of files that can
be processed, probably due to running out of memory.